Доделал функционал редактирования справочника типов приборов учета и добавления нового прибора учета на улицу

+++ src/main/java/model/TypeMeteringDevice.java
@@ -61,7 +61,7 @@ public class TypeMeteringDevice {
         return facility;
     }
 
-    public double getCapacity() {
+    public Integer getCapacity() {
         return capacity.get();
     }
